requirements.txt
click==3.3
mock==1.0.1
pytest==2.6.2
requests==2.4.1
requirements-parser==0.0.6
caniusepython3==2.2.0
